fix(solvable_files): Fix quotes in sql statement of gf-pgsql case
In SQL single quotes must be used to surround a literal string.
The path in this case is a literal string and must be quoted.
The whole CONVERT_FROM statement however is not a literal string.
Therefore it should not be surrounded by single quotes.

In the case for normal users and pgsql the quotes are correct.